+++ packages/core/docs/checkout-com/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.0.7 (not released)
 
 - Removed `ckoWebHookUrl` ([#4910](https://github.com/DivanteLtd/vue-storefront/issues/4910))
-- Storing payment token inside sessionStorage and clearing it in onUnmounted ([#4893](https://github.com/DivanteLtd/vue-storefront/issues/4893))
+- Storing payment token inside sessionStorage and exported `removeTransactionToken` method ([#4928](https://github.com/DivanteLtd/vue-storefront/issues/4928))
 
 ## 0.0.6